Move widgets and utilities to separate folders
