+++ lib/src/utilities/utils.dart
@@ -1,6 +1,6 @@
 import 'dart:math' as math;
 import 'package:flutter/material.dart';
-import 'package:sleek_circular_slider/src/unit_conversions.dart';
+import 'package:sleek_circular_slider/src/utilities/unit_conversions.dart';
 
 bool isPointInsideCircle(
   Offset point,
