+++ lib/src/widgets/slider_label.dart
